Simplify functions in fit_hoi
Figured out that the model fits much better with varying
coefficients for the species as well as varying exponents. This
modification for the annual plant model doubles the number of
parameters for the model and makes finding convergent solutions
much more difficult.

Need to write some functions in fit_hoi to try varying initial
values to find good fit.