Add new mechanistic simulation
New mechanistic simulation *3* is based on sublinear
scaling of resource uptake rates with biomass and
linear scaling of respiration and loss rates with
biomass.  Simulates individual plant growth rather
than total biomass of population.

This new simulation produces dynamics that are closer
to the fit of the Hassel model and so make for a cleaner
demonstration of HOIs.